cli: move subcommands into cmd package
We would like to test the commands programmatically from Go, so we need
to make them importable.
