+++ cli/cmd/assets/coordinator-policy-hash
