+++ cli/cmd/assets/genpolicy
