+++ cli/cmd/assets/genpolicy-rules.rego
