+++ cli/cmd/assets/genpolicy-settings.json
