+++ packages/by-name/nunki/package.nix
@@ -48,9 +48,9 @@ buildGoModule rec {
   subPackages = [ "coordinator" "initializer" "cli" ];
 
   prePatch = ''
-    install -D ${lib.getExe genpolicy} cli/assets/genpolicy
-    install -D ${genpolicy.settings-dev}/genpolicy-settings.json cli/assets/genpolicy-settings.json
-    install -D ${genpolicy.rules}/genpolicy-rules.rego cli/assets/genpolicy-rules.rego
+    install -D ${lib.getExe genpolicy} cli/cmd/assets/genpolicy
+    install -D ${genpolicy.settings-dev}/genpolicy-settings.json cli/cmd/assets/genpolicy-settings.json
+    install -D ${genpolicy.rules}/genpolicy-rules.rego cli/cmd/assets/genpolicy-rules.rego
   '';
 
   CGO_ENABLED = 0;
